feat(ourlogs): Add top n querying
